Refactor ticket controller methods to improve authorization handling and error responses

+++ app/Http/Controllers/Api/V1/AuthorTicketsController.php
@@ -2,94 +2,73 @@
 
 namespace App\Http\Controllers\Api\V1;
 
-use App\Models\Ticket;
-use App\Policies\V1\TicketPolicy;
 use App\Http\Filters\V1\TicketFilter;
-use App\Http\Resources\V1\TicketResource;
+use App\Http\Requests\Api\V1\ReplaceTicketRequest;
 use App\Http\Requests\Api\V1\StoreTicketRequest;
 use App\Http\Requests\Api\V1\UpdateTicketRequest;
-use App\Http\Requests\Api\V1\ReplaceTicketRequest;
-use Illuminate\Auth\Access\AuthorizationException;
-use Illuminate\Database\Eloquent\ModelNotFoundException;
+use App\Http\Resources\V1\TicketResource;
+use App\Models\Ticket;
+use App\Models\User;
+use App\Policies\V1\TicketPolicy;
 
 class AuthorTicketsController extends ApiController
 {
     protected $policyClass = TicketPolicy::class;
 
-    public function index($author_id, TicketFilter $filters)
+    public function index(User $author, TicketFilter $filters)
     {
         return TicketResource::collection(
-            Ticket::where('user_id', $author_id)->filter($filters)->paginate()
+            Ticket::where('user_id', $author->id)->filter($filters)->paginate()
         );
     }
 
-    public function store(StoreTicketRequest $request, $author_id)
+    /**
+     * Store a newly created resource in storage.
+     */
+    public function store(StoreTicketRequest $request, User $author)
     {
-        try {
-            $this->isAble('store', Ticket::class);
-
+        if ($this->isAble('store', Ticket::class)) {
             return new TicketResource(Ticket::create($request->mappedAttributes([
-                'author' => 'user_id',
+                'author' => $author->id
             ])));
-        } catch (AuthorizationException $ex) {
-
         }
+
+        return $this->notAuthorized('You are not authorized to create that resource');
     }
 
-    public function replace(ReplaceTicketRequest $request, $author_id, $ticket_id)
+    public function replace(ReplaceTicketRequest $request, User $author, Ticket $ticket)
     {
-        try {
-            $ticket = Ticket::where('id', $ticket_id)
-                ->where('user_id', $author_id)
-                ->firstOrFail();
-
-            $this->isAble('replace', Ticket::class);
-
+        // PUT
+        if ($this->isAble('replace', $ticket)) {
             $ticket->update($request->mappedAttributes());
-
             return new TicketResource($ticket);
-        } catch (ModelNotFoundException $e) {
-            return $this->error('Ticket cannot be found.', 404);
-        } catch (AuthorizationException $e) {
-            return $this->error('You are not authorized to update that resource', 401);
         }
+
+        return $this->notAuthorized('You are not authorized to update that resource');
+
     }
 
-    public function update(UpdateTicketRequest $request, $author_id, $ticket_id)
+    public function update(UpdateTicketRequest $request, User $author, Ticket $ticket)
     {
-        try {
-            $ticket = Ticket::where('id', $ticket_id)
-                ->where('user_id', $author_id)
-                ->firstOrFail();
-
-            $this->isAble('update', Ticket::class);
-
+        // PUT
+        if ($this->isAble('update', $ticket)) {
             $ticket->update($request->mappedAttributes());
-
             return new TicketResource($ticket);
-        } catch (ModelNotFoundException $e) {
-            return $this->error('Ticket cannot be found.', 404);
-        } catch (AuthorizationException $e) {
-            return $this->error('You are not authorized to update that resource', 401);
         }
+
+        return $this->notAuthorized('You are not authorized to update that resource');
     }
 
-    public function destroy($author_id, $ticket_id)
+    /**
+     * Remove the specified resource from storage.
+     */
+    public function destroy(User $author, Ticket $ticket)
     {
-        try {
-            $ticket = Ticket::where('id', $ticket_id)
-                ->where('user_id', $author_id)
-                ->firstOrFail();
-
-            $this->isAble('delete', Ticket::class);
-
+        if ($this->isAble('delete', $ticket)) {
             $ticket->delete();
-
             return $this->ok('Ticket successfully deleted');
-        } catch (ModelNotFoundException $e) {
-            return $this->error('Ticket cannot be found', 404);
-        } catch (AuthorizationException $e) {
-            return $this->error('You are not authorized to update that resource', 401);
         }
+
+        return $this->notAuthorized('You are not authorized to delete that resource');
     }
 }
